+++ opentelemetry-sdk/tests/metrics/test_measurement_consumer.py
@@ -39,7 +39,6 @@ class TestSynchronousMeasurementConsumer(TestCase):
                 resource=Mock(),
                 metric_readers=reader_mocks,
                 views=Mock(),
-                enable_default_view=Mock(),
             )
         )
         self.assertEqual(len(MockMetricReaderStorage.mock_calls), 5)
@@ -56,7 +55,6 @@ class TestSynchronousMeasurementConsumer(TestCase):
                 resource=Mock(),
                 metric_readers=reader_mocks,
                 views=Mock(),
-                enable_default_view=Mock(),
             )
         )
         measurement_mock = Mock()
@@ -78,7 +76,6 @@ class TestSynchronousMeasurementConsumer(TestCase):
                 resource=Mock(),
                 metric_readers=reader_mocks,
                 views=Mock(),
-                enable_default_view=Mock(),
             )
         )
         for r_mock, rs_mock in zip(reader_mocks, reader_storage_mocks):
@@ -99,7 +96,6 @@ class TestSynchronousMeasurementConsumer(TestCase):
                 resource=Mock(),
                 metric_readers=[reader_mock],
                 views=Mock(),
-                enable_default_view=Mock(),
             )
         )
         async_instrument_mocks = [MagicMock() for _ in range(5)]